feat: implement sign in page (#7)
