+++ index.html
@@ -10,6 +10,11 @@
       as="style"
       crossorigin
     />
+    <link
+      href="https://cdn.jsdelivr.net/gh/orioncactus/pretendard@v1.3.9/dist/web/static/pretendard.min.css"
+      as="style"
+      crossorigin
+    />
     <!-- Favicon -->
     <link
       rel="icon"
