+++ src/entities/index.ts
@@ -6,3 +6,4 @@ export { Button, DisableButton, TwoOptionsButton } from "./element/Button";
 export { Background } from "./screen/Background";
 export { GrayBackground } from "./screen/GrayBackground";
 export { Loading } from "./screen/Loading";
+export { Splash } from "./screen/Splash";
