+++ src/shared/configs/axios.ts
@@ -5,7 +5,6 @@ export const API = axios.create({
   baseURL: import.meta.env.VITE_SERVER_URL,
   headers: {
     "Content-Type": "application/json",
-    "ngrok-skip-browser-warning": "69420",
   },
 });
 
@@ -13,7 +12,6 @@ export const FORMAPI = axios.create({
   baseURL: import.meta.env.VITE_SERVER_URL,
   headers: {
     "Content-Type": "multipart/form-data",
-    "ngrok-skip-browser-warning": "69420",
   },
 });
